# dormitory-backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dormitory API",
    "description": "API for managing dormitory operations including authentication, admin, areas, registrations, reports, rooms, users, contracts, bills, notifications, services, and payments",
    "version": "1.0.0"
  },
  "paths": {
...
   "/api/services": {
      "get": {
        "tags": ["Service"],
        "summary": "Get all services (Admin)",
        "description": "Retrieve a paginated list of all services. Requires admin privileges.",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "Page number"},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}, "description": "Items per page"}
        ],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {"type": "array", "items": {"$ref": "#/components/schemas/Service"}},
                    "total": {"type": "integer"},
                    "pages": {"type": "integer"},
                    "current_page": {"type": "integer"}
                  }
                }
              }
            }
          },
          "403": {"description": "Admin access required"}
        }
      },
      "post": {
        "tags": ["Service"],
        "summary": "Create a new service (Admin)",
        "description": "Create a new service. Requires admin privileges.",
        "security": [{"Bearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string", "example": "Điện"},
                  "unit": {"type": "string", "example": "kWh"}
                },
                "required": ["name", "unit"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Service"}
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "Tên dịch vụ không được để trống",
                        "Đơn vị không được để trống",
                        "Tên dịch vụ đã tồn tại",
                        "Tên dịch vụ không được vượt quá 100 ký tự",
                        "Đơn vị không được vượt quá 10 ký tự",
                        "Dữ liệu JSON không hợp lệ"
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {"description": "Admin access required"}
        }
      }
    },
    "/api/services/{service_id}": {
      "get": {
        "tags": ["Service"],
        "summary": "Get service by ID (Admin)",
        "description": "Retrieve details of a service by ID. Requires admin privileges.",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "service_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Service ID"}
        ],
        "responses": {
          "200": {
            "description": "Service details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Service"}
              }
            }
          },
          "404": {"description": "Service not found"},
          "403": {"description": "Admin access required"}
        }
      },
      "put": {
        "tags": ["Service"],
        "summary": "Update service (Admin)",
        "description": "Update an existing service. Requires admin privileges.",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "service_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Service ID"}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string", "example": "Nước"},
                  "unit": {"type": "string", "example": "m3"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Service"}
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "Tên dịch vụ không được để trống",
                        "Đơn vị không được để trống",
                        "Tên dịch vụ đã tồn tại",
                        "Tên dịch vụ không được vượt quá 100 ký tự",
                        "Đơn vị không được vượt quá 10 ký tự",
                        "Dữ liệu JSON không hợp lệ"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {"description": "Service not found"},
          "403": {"description": "Admin access required"}
        }
      },
      "delete": {
        "tags": ["Service"],
        "summary": "Delete service (Admin)",
        "description": "Delete a service by ID. Requires admin privileges.",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "service_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Service ID"}
        ],
        "responses": {
          "204": {"description": "Service deleted"},
          "400": {
            "description": "Service has related rates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Không thể xóa dịch vụ vì có mức giá liên quan. Vui lòng xóa các mức giá trước."
                    }
                  }
                }
              }
            }
          },
          "404": {"description": "Service not found"},
          "403": {"description": "Admin access required"}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "service_id": {"type": "integer"},
          "name": {"type": "string"},
          "description": {"type": "string"}
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}